refactor(appbars): Apply new compose contract convention.
