+++ appbars/src/main/java/com/decathlon/vitamin/compose/appbars/OverflowMenu.kt
@@ -5,14 +5,16 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import com.decathlon.vitamin.compose.appbars.topbars.ActionItem
+import com.decathlon.vitamin.compose.appbars.topbars.icons.VitaminMenuIconButtons
 import com.decathlon.vitamin.compose.dropdown.VitaminMenus
 
 @Composable
 internal fun OverflowMenu(
     actions: List<ActionItem>,
     modifier: Modifier = Modifier,
-    onDismissRequest: (() -> Unit)? = null,
     expanded: MutableState<Boolean> = remember { mutableStateOf(false) },
+    onDismissRequest: (() -> Unit)? = null,
     overflowIcon: @Composable VitaminMenuIconButtons.() -> Unit
 ) {
     VitaminMenus.Dropdown(
