+++ appbars/src/main/java/com/decathlon/vitamin/compose/appbars/bottomnavigations/BottomNavigationColors.kt
@@ -1,0 +1,11 @@
+package com.decathlon.vitamin.compose.appbars.bottomnavigations
+
+import androidx.compose.runtime.Immutable
+import androidx.compose.ui.graphics.Color
+
+@Immutable
+class BottomNavigationColors(
+    val background: Color,
+    val selected: Color,
+    val unSelected: Color
+)
