+++ appbars/src/main/java/com/decathlon/vitamin/compose/appbars/topbars/icons/VitaminMenuIconButtons.kt
@@ -1,0 +1,33 @@
+package com.decathlon.vitamin.compose.appbars.topbars.icons
+
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.material.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import com.decathlon.vitamin.compose.appbars.R
+
+object VitaminMenuIconButtons {
+    @Composable
+    fun More(
+        onClick: () -> Unit,
+        contentDescription: String?,
+        modifier: Modifier = Modifier,
+        enabled: Boolean = true,
+        interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
+    ) {
+        IconButtons(
+            onClick = onClick,
+            modifier = modifier,
+            enabled = enabled,
+            interactionSource = interactionSource,
+            icon = {
+                Icon(
+                    painter = painterResource(R.drawable.ic_vtmn_more_2_line),
+                    contentDescription = contentDescription
+                )
+            }
+        )
+    }
+}
